Extract JavaScript format and lint rules
Similar to before,
churn while changing the JS rules means constant rebuilding.
Not the best thing when that means re-installing packages from scratch.
